updated/rearranged courses section
Removed MyBatis/Jsp udemy course due to the preparations for the final push in the ENTRA21 Bootcamp,  Moved GitHub and OCA-808 course to the 'Concluídos' section,  Added Spring Boot course to the 'Em Progresso' section. Fixed missing links from courses, Improved the introduction on the top of the document, linked the Hibernate project.